refactored style_new.py and fully deprecated style.py
